requirements.txt
streamlit==1.43.2
google-generativeai==0.8.4
pytube==15.0.0
python-dotenv==1.0.0
Pillow==11.1.0
pydantic==2.10.6
toml==0.10.2
tqdm==4.67.1 